requirements.txt
anyio==4.2.0
attrs==25.3.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.6
colorama==0.4.6
command-not-found==0.3
dbus-python==1.3.2
distro-info==1.7+build1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
instaloader==4.14.1
markdown-it-py==3.0.0
mdurl==0.1.2
nala==0.15.1
netifaces==0.11.0
outcome==1.3.0.post0
pexpect==4.9.0
praw==7.8.1
prawcore==2.4.0
ptyprocess==0.7.0
Pygments==2.17.2
PyGObject==3.48.2
PySocks==1.7.1
python-apt==2.7.7+ubuntu4
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
selenium==4.31.0
setuptools==68.1.2
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.9.0
typing_extensions==4.10.0
ubuntu-pro-client==8001
ufw==0.36.2
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
wheel==0.42.0
wsproto==1.2.0
